Utilities fonction (#9)
Improve plot for Neurips.
We double the torch precision because for synthetic data, the loss is sometimes set to 0 while it should be 10^-12.
Dispatch evaluation based on model and data_loader type (to compute accuracy/loss).
Training settings.
Test on Pytorch utilities functions.
Refactoring. Put all out-of-date in a dedicated package (might be interesting for future work).
Remove unecessary Nets.